LayoutTests/imported/w3c/web-platform-tests/css/css-values/exp-log-serialize-expected.txt


FAIL 'exp(0)' as a specified value should serialize as 'calc(1)'. assert_equals: 'exp(0)' and 'calc(1)' should serialize the same in specified values. expected "calc(1)" but got "exp(0)"
FAIL 'scale(exp(0))' as a specified value should serialize as 'scale(calc(1))'. assert_equals: 'scale(exp(0))' and 'scale(calc(1))' should serialize the same in specified values. expected "scale(calc(1))" but got "scale(exp(0))"
PASS 'exp(0)' as a computed value should serialize as '1'.
PASS 'scale(exp(0))' as a computed value should serialize as 'matrix(1, 0, 0, 1, 0, 0)'.
FAIL 'log(1)' as a specified value should serialize as 'calc(0)'. assert_equals: 'log(1)' and 'calc(0)' should serialize the same in specified values. expected "calc(0)" but got "log(1)"
FAIL 'scale(log(1))' as a specified value should serialize as 'scale(calc(0))'. assert_equals: 'scale(log(1))' and 'scale(calc(0))' should serialize the same in specified values. expected "scale(calc(0))" but got "scale(log(1))"
PASS 'log(1)' as a computed value should serialize as '0'.
PASS 'scale(log(1))' as a computed value should serialize as 'matrix(0, 0, 0, 0, 0, 0)'.
PASS 'calc(exp(0) + log(1) + log(1))' as a specified value should serialize as 'calc(1)'.
PASS 'scale(calc(exp(0) + log(1) + log(1)))' as a specified value should serialize as 'scale(calc(1))'.
PASS 'calc(exp(0) + log(1) + log(1))' as a computed value should serialize as '1'.
PASS 'scale(calc(exp(0) + log(1) + log(1)))' as a computed value should serialize as 'matrix(1, 0, 0, 1, 0, 0)'.
PASS 'calc(log(1) + 0.5)' as a specified value should serialize as 'calc(0.5)'.
PASS 'scale(calc(log(1) + 0.5))' as a specified value should serialize as 'scale(calc(0.5))'.
PASS 'calc(log(1) + 0.5)' as a computed value should serialize as '0.5'.
PASS 'scale(calc(log(1) + 0.5))' as a computed value should serialize as 'matrix(0.5, 0, 0, 0.5, 0, 0)'.
PASS 'calc(log(0))' as a specified value should serialize as 'calc(-infinity)'.
PASS 'scale(calc(log(0)))' as a specified value should serialize as 'scale(calc(-infinity))'.
FAIL 'calc(log(0))' as a computed value should serialize as '-infinity'. assert_equals: '-infinity' should round-trip exactly in computed values. expected "-infinity" but got "1"
FAIL 'scale(calc(log(0)))' as a computed value should serialize as 'matrix(-infinity, 0, 0, -infinity, 0, 0)'. assert_equals: 'matrix(-infinity, 0, 0, -infinity, 0, 0)' should round-trip exactly in computed values. expected "matrix(-infinity, 0, 0, -infinity, 0, 0)" but got "none"
